refactor: move client-facing files to new dir
