+++ src/client/eagerLoader.js
@@ -1,4 +1,4 @@
-import renderComponent from './_renderer'
+import renderComponent from './renderer'
 import { toMountPointById } from './toMountPointById'
 
 export default function eagerLoader({ id, Component = () => null, props = {} }) {
