+++ src/client/lazyLoader.js
@@ -11,7 +11,7 @@ export default function lazyLoader({ id, componentImporter = () => () => null, p
       if (entry.isIntersecting) {
         const mountPoint = toMountPointById(id)
         if (!mountPoint.getAttribute('data-s-is-hydrated')) {
-          const { default: renderComponent } = await import('./_renderer')
+          const { default: renderComponent } = await import('./renderer')
           const { default: Component } = await componentImporter()
           renderComponent({ mountPoint, Component, props })
           mountPoint.setAttribute('data-s-is-hydrated', true)
