+++ src/client/renderer.js
