+++ src/client/toMountPointById.js
@@ -1,0 +1,3 @@
+export function toMountPointById(id) {
+  return document.querySelector(`slinkity-react-mount-point[data-s-id="${id}"]`)
+}
